added record_to_file function
record to file does not implement any filtering but it also does not use any space on the heap, will need to be modified to use a small amount of heap space for filtering. this is nice because it can run forever but it only seems to be recording about 16 samples / second, will need to determine if this is sufficient and if not figure out how to speed it up.