added parsing of data recorded to file
implemented string parsing of accel and gyro values that have been recorded to file and send it to firebase. still having trouble doing firebase sends greater than ~10 samples, we'll have to figure out a way to send in bursts. ideally we read the file with recordings in bursts and send to firebase in bursts but this will require a modification of how we are currently converting the tables into jsons.