fixed sending to firebase
reorganized way we send to firebase to utilize callback function from http_post correctly. instead of calling the function to read the next line directly from the callback, instead need to use
node.task.post(send_line)
to post the function as a task to be run, this forces send_line() function to wait for the http post to get a return before executing the next one.